requirements.txt
Flask
flask-cors
python-dotenv
numpy
opencv-python
tensorflow
requests
Pillow
Werkzeug
